repository added

+++ example/java/kotlin/provider/src/main/kotlin/br/com/zup/pact/provider/dto/BalanceDTO.kt
@@ -5,5 +5,5 @@ import java.math.BigDecimal
 data class BalanceDTO(
         val clientId: Int,
         val accountId: Int,
-        val balance: BigDecimal
+        val balance: Double
 )
